Adds global database variable to db subpackage
Apparently, this is intended to be kept open for the lifetime of the
application. It maintains a connection pool on its own and can be passed
around to goroutines and the like.

Also defines package-level helpers for getting a transaction (single
connection from the pool), and opening/closing the global database pool.